Change eventloop.WaitOnRegistered to execute callbacks
This is a behaviour change which will make it possible for users of the
eventloop to use it as the only way to stop stuff, without needing to do
additional stuff if the event loop has errored out, detect it and have
a separate way to stop without using it.

The above is in practice very error prone and neigh impossible.

In the future when we have additional mechanism for notifying users of
an iteration ending, this behaviour will stay the same as using the
event loop is still likely the best way to synchronise around stuff such
as the runtime.