Fix panic: GoError: creating a new page: canceled
When we run multi-VU/instance tests, one of the test runs can close a
page of its own. Since each test run attaches to all the browser pages
(this is another area we might want to discuss), this sometimes
creates a race between the test runs. This check allows the test run
to continue instead of panicking because of another test run's page
detachments.

This fix allows running the k6 browser in high concurrency with
multiple VUs and instances. Note that it is still possible to get
errors since we need to correctly handle the CDP messages while
sending and receiving them (the order of them). This can cause
timeout and other sorts of errors.

Explanation:
There are two instances and different sessions. However, both
instances attach/detach from the same pages since we run in the same browser.

instance1 ---> newPage                    page1
  attaches to: page1
instance2
  attaches to: page1
instance2 ---> newPage                    page1, page2
  attaches to: page2
instance1
  attaches to: page2
instance1 --> page1.Close()               page2
instance1 <-- detachedFromTarget(page1)
  closes page1 session.

This is the racy part that this PR fixes:

instance2 <-- detachedFromTarget(page1)
  closes page1 session.
instance2 ---> attachToTarget(page1)
  panic: session does not exist.

instance2 panics while trying to send CDP messages to a closed web socket
connection since it detaches from it and tries to attach to it
concurrently.