Fix TestVUIntegrationMetrics not checking metric exactly
turns out the test has been *slightly* broken as it checks the metric
against a specific builtinMetrics instance.

Unfortunately that was the instance for the "archive" test and the
"source" one was never used.

This works as Equals will compare recursively and both the registry and
the builtinMetrics should be the same.

I only noticed it as a test failed with this, but it likely was supposed
to fail with something differently. For example a sink having
a different value :shrug: