Remove browser_first_paint metric
This metric isn't useful as it only helps us work out how long it took
to render the first element on the page, which is usually the
background color. With web vitals a more useful metric is already
exposed and tells us more on how performant and useable a website is
(webvital_first_contentful_paint).