Fix wait for detached selector
This commit fixes waiting for a selector to be in 'detached' state.
Fixes the injected_script wait for selector condition for 'detached'
state, that was returning 'undefined' which eventually caused an
infinite loop from the calling polling function
'waitForPredicateFunction' due to the call:
    return predicateFn(...args) || continuePolling;
As 'undefined' returned from the predicate function resulted in
returning 'continuePolling' from this condition.

Also modifies the Locator.waitFor method in order to call the
Frame.waitFor method, instead of Frame.waitForSelector, as in this case,
we do not need to retrieve the associated ElementHandle for the
selector.