k6/execution: Add metadata JS API
A bare minimum API to set and get metadata set for the whole VU.

It is added on top of the vu property and on top of a new `metrics` one
so. If `k6/execution` is imported as `exec` -
`exec.vu.metrics.metadata["foo"]` will get you the currently set
metadata for `foo` the same way `exec.vu.tags["bar"]` will get you the
value of the tag `bar`.

Setting metadata is the same.

It also adds `exec.vu.metrics.tags` which is the same as `exec.vu.tags`
for consistency. While leaving the old one for backwards compatibility.

This also includes dropping the ability to list metadata keys and values
from `exec.vu.tags`. This was likely left in by mistake.

This is a breaking change due to the fact that before that `iter` and
`vu` could've been accessed this way but this will now need to be done
through the new `exec.vu.metrics.metadata`.

Closes #2766