Make scenario options JSON field optional
Previously, the new scenario `options` field added in #3036 was being
serialized to JSON by default, which made it an incompatible change with
some of our internal Cloud services that validate the existing
configuration structure.

This change makes the object optional, and it will only be serialized if
it's included in the data.