Disable timeout on browser process execution (#866)
* Disable timeout on browser process launch

This commit modifies the browser launching process to not pass a context
with timeout to the actual browser process cmd handle. Currently this
was done using as timeout the value set for the browser.timeout option,
which should serve a different purpose.

* Disable timeout on browser process connection

This commit modifies the browser connecting process to not pass a
context with timeout to the browser process implementation so its
closing handler is not triggered after the set timeout. Currently this
was done using as timeout the value set for the browser.timeout option,
which should serve a different purpose.