Fix testbrowser custom ctx does not wrap VU state
Test browser had support for withContext option, which allowed to pass a
custom context to test browser, which would be the one set as VU
context. Usually this context was used in order to control from the test
the cancellation of the test browser.
The problem is that this context passed from the test implementation,
was "divergent" from the test runtime context set up when calling
setupHTTPTestModuleInstance() from test browser, which sets up all the
VU related mock data, including the mandatory browser type field inside
scenario options. Because this option is not mandatory, tests that were
using a custom context were failing due to scenario options not being
present in this context.

This has been fixed by not allowing custom contexts to be set for test
browser, and instead expose this component's context and context
cancelling function so they can be handled from tests implementation.

This commit also fixes the tests affected by this issue.