Initialize browser implicitly
Because browser is not controlled by explicit JS test code, now every
browser level method will automatically initialize the browser for the
iteration in case it has not been initialized yet in a previous call.
Once the browser is initialized it's stored in the browser registry
indexed by the iteration ID (composed by VUID-scenario-iterationID).
When the iteration ends, signaled by the VU context being done, the
browser is automatically closed and removed from the registry.