Fix random ws url selection
When K6_BROWSER_WS_URL is a comma separated list of ws urls, we need
to pick from the list every time remoteRegistry.IsRemoteBrowser is
called, not just once when remoteRegistry is created. This change moves
the round robin (with the use of rand) to remoteRegistry.Is...
...RemoteBrowser from env.IsRemoteBrowser. env.IsRemoteBrowser now
only collects and returns the list of ws urls.