Use vu.Context when pushing metrics
Previously, the browser implementation components, such as browser,
page, frame... were holding a context inherited from the vu.context
because the browser lifecycle was tight to it. Therefore, we were
passing these contexts when calling k6metrics.PushIfNotDone, as if that
context was done, it meant that the VU context was also done and at the
same time the VU metrics samples channel was closed. With the current
implementation, because the browser, page, frame... contexts do not
inherit from the VU.context, we can not use these contexts when calling
k6metrics.PushIfNotDone, as it is not a guarantee that the vu samples
channel is open, so we should use the VU context directly.