Add a method to parse and create remoteRegistry
This new method parses K6_BROWSER_WS_URL differently. Instead of a list
of ws urls, it is a scenarios object, which needs to be unmarshalled
and all ws urls extracted. The end result is the same -- a list of ws
urls.

This is a step closer to being able to work with a set of chrome
browsers each with different characteristics that are set in the
scenario.

This commit contains the addition of this new method. Eventually this
method will be used instead of the existing newRemoteRegistry method.

Closes: https://github.com/grafana/xk6-browser/issues/888