Bucket start point as flushed time
Use the start point of the bucket as the flushed time instead of the center of the bucket.

Before we used the center for doing the same as we did on the version 1.
The rationale behind it are not relevant anymore for the version 2 so we
simplified it.
Version 1 used it based on the assumption that there is a uniform distribution
of the samples in the bucket so the center is a more accurate distance
as a generic value. This was relevant because we used to push samples
aggregated and not (the outliers). We now flushes only aggregated
values.