Correct stop on error signal
The previous version using the PeriodicFlusher helper was not correct.
On Stop the PeriodicFlusher triggers again the callback so before
really stopping we was calling again the flush.

The new architecture does the same as the PeriodicFlusher but it does
not call again on stop. The flush on correct Stop is handled directly
from the stop method.