Move init of remoteRegistration
Since creating a remoteRegistry now errors, we need to be able to
send a signal back to k6 to let it know to shut down cleanly. At the
moment there is no way to do this through New. We can achieve this
with a workaround in NewModuleInstance though. Since NewModuleInstance
is called once per vu and remoteRegistry only needs to be created
once, we have to use sync.Once to init a new remoteRegistry once
which is then used by all vus.