Remove some of the http metrics
The motivation for this came from analysing and working on removing the
web vital rating metrics. It became apparent that we didn't need to
expose all these http metrics just yet and instead should focus on the
ones that are most used. We used k6 for guidance on this and they have
suggested that browser_http_req_duration and browser_http_req_failed
are the most useful. We can add metrics back in when users show
interest in them. This will for now also reduce the overall number of
time series metrics.