Revert back to working with the defaultContext
When multiple k6 instances connect to a single chrome instance, the
page that is created by each k6 instance is also recreated in all
k6 instances that are connected to the same browser. We don't do the
same for browserContexts. When a page is recreated in one k6 instance
for another k6 instance, the browserContext will not be there, so
instead for now we should still work with the defaultContext.