Wait for IterStart and IterEnd events to be processed
The browser module uses the IterStart and IterEnd events for browser
initialization and shutdown, so we need to wait for them to complete.
There is some concern that this might add some delay to the overall
iteration duration, and not just to the iterations where browsers
processes are managed, but it should be negligible overall. We should
run benchmarks after this change to confirm this.