Fix CI integration issue by upgrading runner
We're upgrading the runner to ubuntu-latest-8-cores. We've
experimented with in other investigative PRs, and it has helped
resolve the issue where the integration tests are failing in the CI
pipeline.

This is a short term fix, in the hopes that we can find a more
long term fix by working closely with github, or watching out for
their next release of the runner.