Fix warn level log to debug
While testing the fill form example script against a remote chrome
instance, it became apparent that this warning wasn't useful
information for the end user. Downgrading from warn to info should
avoid noise in the logs.

All other areas which work with fetchBody and receive an error will
remain to behave as they currently do, which is to assume an error,
which still represents the expected behaviour for those actions.