Remove goroutine ID field from logger
Adding the goroutine ID as a log field required copying the calling
goroutine stack into a buffer and parsing the ID from it. This was done
for each log action which is very inefficient and does not provide much
value as we are already using a field for the iteration ID. Therefore
our decision is to remove it completely.