Refactor onConsoleAPICalled log fields
In order to be consistent with the usage of 'source:browser' log fields
across all logs generated from the browser extension, this commit
replaces the current 'source' log field for onConsoleAPICalled logger,
which was previously set to 'browser-console-api', to 'browser'. Also
moves and modifies the previous 'browser-console-api' to be set as
'console-api' for the 'browser_source' keyed field, which is currently
also used in the 'onLogEntryAdded' method logger.