Avoid onRequestPaused error log when ctx canceled
Continuing/Failing a request from onRequestPaused happens when CDP's
network.fetch functionality is enabled and therefore k6 browser acts as
a proxy for every request that matches the specified pattern.

A context.Canceled error here happens most probably because the
iteration has ended, closing the browser's context inherited by
NetworkManager, meanwhile the CDP command is being executed.

Given this situation, and considering that this happens often in sites
that are continuously making background requests, showing these error
messages is misleading for the end user. Therefore is probably a better
solution to change the log level to debug when this situation occurs.