Update panic to log when cdp close fails
While testing for another change, it became apparent that the panic
when working with browser.close() would prevent any of the remaining
data directories from being cleaned up in the temporary directory.
Instead of panicking, we're going to log an error and carry on with
rest of the process to try and close the subprocess and eventually
delete the data directory. This change does indeed bring about the
behaviour we want even when the context has been closed.