Add a test to ensure data dir cleaned up
When the context is closed before we call browser.close() used to
cause a panic on the cdp command to close the browser. Since we've
changed this to log instead of panic, we just need to make sure that
the data dir does indeed get removed when the cdp command fails.