Add skip tip and windows e2e test run
The e2e workflow for windows using the tip version of go doesn't seem
to work on Windows (or at least in the ci github action). The tip is
installed correctly, but when it comes to building the k6 binary using
xk6, xk6 seems to always execute the stable version of the go binary
and not the tip version. When running commands outside of xk6 to
verify the env vars (e.g. GOROOT) it always returns the tip paths. It's
unclear why this is happening.

For now, we're skipping the windows e2e tip workflow.