tc39: Don't run some unicode tests as they are flaky between go versions
Go 1.21 adds support for more unicode, which is great, but this also
means that now some tests are passing that previously weren't on 1.20.

It is preferable that our tests don't fail or not fail between different
versions of go - especially ones that we support.

This can be dropped once 1.21 is the lowest go version we support.