Add core to inject k6 object into window
In some cases we want to pass on information to other apps to allow
them to infer who or what is performing requests against a website.
For now we're exposing an empty object, but we hope to extend the
object with meta data that will be useful for other apps that have been
instrumented on the website.