Experimental implementation of a reference counted `file.close` method
The main issue with this implementation is that as k6 does not yet
support opening files in the VU context, the ref count for a given
file will never reach zero as the VU 0 (init context) holds
one reference until the end of the test.

The whole idea behind reference counted closing was that in a context
with multiple scenarios, once all the instances of a file were explicitly
closed by the user, we would be able to release the memory, regardless
of what over workloads might be happening next.