Add a test to ensure that timeout is propagated
This tests ensures that the timeout that is set in page is propagated
to the underlying objects (frame manager and frame).

There is one issue which is that the timeout is set in milliseconds but
later it is translated into a second. This needs to be resolved in a
new PR.