Update goja
This updates makes the runtime initialization faster and *lighter*.

For empty and fairly small scripts the observed initialization
improvement by some fast "benchmarking" for k6 this means:
1. around 3x faster initialization of VUs
2. around 4x less memory usage.

This likely will be nearly unnoticeable as this is only reasonably easy
to see with 100k VUs for the initialization speed where it goes from
around 10s to a little over 3s.

Adding more code increases both linearly, so bigger scripts will have
some speed up but it won't be 3x.

Same seems to be for memory usages as well.

Both of those likely get way worse if more and more of the JavaScript
"standard library" is used as in that case all of those now templated
properties will be used.