Add no-sandbox arg to browser enabled Docker image
In order to allow chrome to execute in the browser enabled image, either
SYS_ADMIN Docker capability or no-sandbox browser argument had to be
set. This commit sets the 'no-sandbox' option in the Dockerfile
definition itself so this action is no longer necessary to run the
Docker container. This is arguably also the better option, as we are
running the chrome browser inside a container, versus using the Docker
capability.