Fix browserContext.waitForEvent's timeout
This now defaults to the browserContext's default timeout. This should
fix the confusion of mixing seconds and milliseconds when working with
the timeouts in this API.

NOTE though that there is a default timeout of 30 seconds now, whereas
in PW there is no timeout unless one is set explicitly in
browserContext using SetDefaultTimeout.