Fix race in http.batch after latest goja update
Previous http.batch code did call `runtime.NewArrayBuffer` concurrently.

Which never has been guaranteed to be safe, but with the latest goja
changes to having prototypes created on demand it races.

The fix *might* have some performance implications, but they are likely
to be very small as `NewArrayBuffer` mostly wraps stuff so ... hopefully
not a big deal.

Fixes #3345