Refactor browserContext.waitForEvent tests
This refactors the tests so that it matches the new signature for
waitForEvent, which is that it takes in a predicate method and a
timeout.

The predicate method will now be wrapped and added to a taskqueue in
the mapping layer which means we do not need to work with the event
loop in the test itself.

Since the options are now parsed in the mapping layer, this also
simplifies how we pass the predicate and timeout arguments to
waitForEvent.