Fix taskqueue so it is per page
The taskqueue wasn't working when it was one per vu since it became
difficult to understand exactly when it was best to close the taskqueue
if there were multiple pages.

This change now creates a taskqueue per page, but the taskqueues is
stored in the vu rather than in the page struct itself, which helps
ensure that the taskqueue is kept away from the page business logic.