Add tracing helper functions
This functions will try to call the same method they implement from the
tracer contained in the given context. If the tracer is not found in the
context, they return a noop Span. This makes the code in the common
package cleaner and avoids dealing with possible NPEs. At the sime time
avoids having to mock the tracer in tests.