Move log tokenizer to lib/parseutils package
Up until now the tokenizer implementation was only used to parse the log
output option configuration. With our current intention of using this
format for other options, such as traces output, it becomes evident that
this implementation should be moved away from the log package and into a
more generic one.