breaking change: Don't try to prepend https to module specifiers
This has been deprecated for years and never really made much sense, but
made a bunch of the error messages way more convoluted.

Closes #3287

This also fixes tc39 tests and some archive ones that were depending on
this functionality.

AFAIK all of those tests should've not been written the way they were to
begin with - they just happened to work due to the previous behaviour.