Move TracerProvider init/stop to run cmd
Because traces will only be generated during a test run, move the tracer
provider initialization and stop to the run command.
Additionally, for this same reason, there is no need to wait for the
tracer provider to shutdown unless the actual test run is executed,
therefore wait for it concurrently with metrics processing wait.