Fix isVisible for frame and page
isVisible is changed to work with the new runActionOnSelector method,
which does not wait for any elements to match the given selector. It
will just return false if there are no matching elements.

This also means that timeout is an obsolete option, which means that
the test in locator_test is not valid and has been removed.