Add built-in root certificate fallbacks (#3440)
* Add built-in root certificate fallbacks

Blank-importing golang.org/x/crypto/x509roots/fallback bundles a set of
root fallback certificates from Mozilla into the resulting binary. This
allows the program to run in environments where the system root
certificates are not available, for example, inside a minimal container.
These are _fallbacks_, meaning that if the system _does have_ a set of
root certificates, those will be given priority. The binary size will
increase a little (~ 220 kB).

It should be added to main.go instead of somewhere else because the
recommendation is for the package to be imported from binaries, not from
libraries. Calling x509.SetFallbackRoots (what the imported package does
in its init function) more than once will cause the program to panic. In
principle, the Go import system will prevent the package from being
imported twice, so that shouldn't be a problem. That said, xk6 has
opinions, and it doesn't want to have an import in main.go, so it's
added to the only package imported from main, "cmd".

---------

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>
Co-authored-by: Oleg Bespalov <oleg.bespalov@grafana.com>