Add integration tests for evaluate parse object
This integration test will evaluate objects which uses the existing
valueFromRemoteObject function in remote_object and not the latest
parseConsoleRemoteObject function.

In this change there are 3 commented out sub tests, 2 of which will be
fixed, and 1 will remain commented out until a fix for the linked issue
has been resolved.