Fix potential nil exception when getting frame
When a part of the browser module needs to retrieve the frame given an
id, it should check whether the frame is nil before using it. This is
not safe, since it's easy to forget to check for nil.

This change avoids the need to check for nil and reinforces a
requirement to check with an addition of a bool that is also returned
to indicate whether the frame was found or not.