Move goja callable to js string
Move the transformation of the goja callable to a string outside of the
promise for frame.waitForFunction. This is to avoid using the non
thread safe goja runtime from multiple goroutines.

This change also includes a temporary change to page.waitForFunction
which will still do the transformation in the promise for now.