Add RemotePersister
This will be used to upload files (screenshots) to a remote location.
It first retrieves the presigned url for the current file, then uploads
the file with the new presigned url.

It is possible to retrieve multiple presigned urls in one request, but
the issue is that we don't know the names of the files before hand
which is a requirement to retrieve the presigned urls.