Refactor LocalPersister.Persist
This refactor will enable us to swap out the LocalFilePersister and
RemotePersister. The Persist method in LocalFilePersist must match what
is a requirement for the RemotePersister, which is the addition of the
context as a param, even though the LocalFilePersister doesn't work
with it.

These two persisters will be interchangeable at a later stage depending
on the environment setup.