Add a check for target is in the same context
This isn't yet functional until we add other mechanics in later commits.
For now, it allows Connection to attach a target if it's in the same
BrowserContext (or in the default context).

We don't use `defer` for the lock because we don't want introduce any
lock overhead. Local tests showed us that holding the lock for the
entire function is prone to unexpected deadlock issues. This is actually
for a future code addition that can easily miss this fact.