Add nil checks in requestByID
For unknown reasons mainFrameSession or mainFrameSession.networkManager
(it's unknown exactly which one) are nil, which has caused NPDs. We're
now adding nil checks here to prevent the NPDs, but we're also logging
when these are nil in the hopes that we can identify which component is
nil and under what conditions.