Don't "unset" `open` and `require` after the init context
This was done to prevent them to be used outside of the init context,
except that this wasn't enough as you can save them to another variable.

More fixes have been added on top - but in reality there is no need to
unset them.

If anything this makes it harder for users to figure out what happens as
in the case of `require` it will tell you that you can't call
`undefined`.

For `open` this was handled by setting it to a function that just
reports it can't be used outside the init context. But this is already
the case if it isn't unset and then reset to this function.

So this change makes the whole thing more consistent and doesn't do a
bunch of additional and unnecessary operations.