Set working dir to the current one if test is read from stdin
Usually the working dir for a given script is set based on it's path.

For test from stdin that path is file:///- so it is basically the root
folder. With this change this changes to the directory k6 was ran from.

This means when you do `cat script.js | k6 run -` the working dir will
be the same as if you did `k6 run script.js`.

But there is still no way to fix `cat path/to/somewhere/script.js | k6
run -` as k6 can't know where it got it's stdin from.

fixes #1462