k6/html: disable revive linter
While nice to have it it mostly gives lint errors for not documenting
every export method.

As we haven't fixed those in 6 years I don't think we will ever do it
for the 100+ methods. As such it seems better to just ignore it and
lower the golangci-lint noise.

Unfortunately using "//revive:disable:exported" in the file itself did
not work with golangci-lint.