golangci-lint: enable it for all issues not just new
As now all lint issues are fixed we can enable without that option.

Unfortunately the option only looks at code at and around the places
that was changed and if any of it has new lint errors it will trigger.

That does mean that if you update golangci-lint - it won't showcase new
lint errors it triggers.

Also won't work if you add a bunch of code to the end an already
somewhat long function. As the lint error will be at the beginning of
the function.

Closes #769