Fix NPD when using goto on a second page
This fixes an issue when working with a second tab (i.e. opening a new
page in the same existing browserContext), and navigating to a url. In
fact this issue would arise as soon as any API is used from the new
page instance. This was occurring due to the unmapped browserContext
when getting it from the browser.