Backport changes to js/compiler from the ESM branch
The changes revolve around moving away from compiling to mostly parsing
the provided code.

This will be relevant in ESM branch when only the parsing needs to be
done for native ESM modules.

For now it mostly adds some noise, but will make that PR a lot better.

The changes to tc39 are mostly to make it nicer as well.

The tc39 changes are quite strange as they definitely have *never*
worked and do not work in k6 currently, so I have no idea what is
happening there.

It is very likely some error shadowing has been exposed.