Implement import.meta.resolve()
This implements and tests `import.meta.resolve()` a way to get the path
to a module the same way ESM and/or `require` will do.

This doesn't read or load the file in question from the file.

This also showed that we currently always parse code as ECMAScript
modules. Even when the parsed/compiled file was to be wrapped as
CommonJS.

This likely had no other side effects as all other ESM specific syntax
is also either not implemented in k6 or forces CommonJS wrapping
disabled.

Closes #3856