Add `discardResponseMessage` option for gRPC client (#3820)
Now gRPC unary RPCs respects `discardResponseMessage` option. When the option is used the client still fetches the response body but it doesn't serialize the expected content as JSON. It removes a step that would take significant CPU cycles.

Note that unfortunately, gRPC doesn't support discarding completely the response body so it is resolved using emptypb.Empty. This is the reason why the response body is still loaded in memory.