Fix cloud command not listed in k6 command's help text (#3901)
* Fix `cloud` command not listed in k6 command's help text

When implementing the new cloud command structure, we marked the
k6 cloud command as deprecated by setting the cobra.Cmd Deprecated
option.

By doing so, Cobra took over itself to hide the command all in all from
the help text's output, and users couldn't see the cloud command as
part of the "Available Commands".

This commit removes the Cobra.Cmd Deprecated option usage, and ports
the message to the Long version of the help text.

This commit also rephrases the message to be more accurate. Considering
the cloud command itself is not deprecated, but rather its existing
behavior, we have mentioned the original behavior consisting in taking
a test resource as input is deprecated, specifically.

* Add a root command test asserting the expected commands are available

This commit adds a test asserting the root command help text lists the
expected commands as available.