Add event.Abort to abort in testAbortController
This is a temporary solution to a problem the browser module is facing
when a signal (e.g. SIGTERM) is handled by k6, but is not correctly
handled by the browser module. The reason is due to it not working with
the vu context. More details in the open issue:
https://github.com/grafana/xk6-browser/issues/1410.

The introduction of a temporary abort event will be fired, which the
browser module can intercept and action on immediately.