Add MakeMemMapFs to testutils package
While working on another PR I was looking for a way to quickly
initialize a test Fs. I found in the codebase occurences of a
makeMemMapFs function, which was already duplicated twice in the
codebase.

Instead of adding a third copy, I thought I would instead make testutil
out of it. This is exactly what this commit does.