Refactor to work with a timeout on cdp close req
Since the context that is used is a essentially a background context,
we need to have a timeout in place. There is a chance that the browser
process has already exited but the connection thinks it is still alive.
By using the timeout we're ensuring that the iteration can end, which
allows for the events to be emitted by the k6 event system and so allow
k6 to exit in a timely manner.