Update PostData to work with PostDataEntries
From some of my testing it would seem that PostDataEntries doesn't
contain more than a single entry. I've tried with various types of
request data such as:
- application/json
- multipart/related
- application/x-www-form-urlencoded
- application/form-data

and so far none of them exhibits a change in behaviour where more than
one element in PostDataEntries is occupied.

We can work with the 0th entry of PostDataEntries for now until a need
arises for needing to work with the rest of the slice (i.e. we have a
better idea of when that happens).