Remove page on event name check
This is safe to do so as we will no longer pass a rogue string to this
method. Only the page on event names can be passed to this function.

Unless, of course, somebody converts a string with an incorrect name to
PageOnEventName to call this function. This will easily be catched in
reviews, and nobody would do such a thing. We follow the Go idiom of
being pragmatic instead of being highly defensive. Also, this method is
used by the mapping layer, and there is already an event name check.