Add error handling on page.on(metric) handler
The most important thing to do is to handle the error and ensure that
the page.on metric returns false so that the metric tag isn't update.
This is being defensive, since the context has already closed and the
metric emitter will not emit a metric when the context has closed.

We're logging a debug log here since this information will not benefit
the end user and when this error occurs they can't action on it. It
could be useful for debugging though.