tc39: fix some tests around ESM only syntax
the compiler.Parse had 1 boolean argument used for two different things:
1. signal that the parser should wrap the code in order to support
   commonJS
2. signal that if it not the above it should be compiled as a module

no tc39 tests should be wrapped as commonJS, but some of them
(especially the ones directly called in it) also are not modules ... so
this was breaking some tests.

The compiler package likely should be completely redone, but that
started to touch too many things so, lets first at least fix the tests.

This is needed to make *some* top-level-await tests work.