Update Sobek and fix http.File to not use []byte (#4009)
Sobek updates bring fixes from goja for:
- nesting of destructuring with arrays
- using slices from go code in js code

Also fix for async errors in modules triggering the promise rejection
tracker as well, which will be needed for top-level-await

The changes to handling of slices did break some tests and showcased
that k6 was handling `http.File().data` badly. Previously it was
`[]byte` which it really shouldn't be.

This fix makes certain that the provided input is `string` or
`ArrayBuffer` and keeps it that way until it needs to be written.

This aligns with the documentation as well as removes a source of pure
`[]byte` in js code which there shouldn't be any.

Separate PR should be made to drop being able to use `common.ToBytes`
with `[]byte` as that should never happen and should be 100% a bug.

Also made the go code return a pointer as that just makes sense instead
of a struct that will get copied.