Fix waitForSelector by calling waitFor
The fix is in waitFor, and waitForSelector just needs to call waitFor
to enable the fix.

It basically will auto retry up to 20 times when certain errors are
received from chrome. This can happens when the underlying DOM is
changing due to a navigation, and the expected element that matches the
selector is in the newly navigated DOM.