Update to work with ToObject instead of cast
This is a better and safer way to work with an interface{} type when we
need to retrieve the property values from the object. Since the
underlying type of Default could be anything, including a sobek.Value,
it's better to retrieve it as a sobek.Object instead of a casting.