Update frameNavigated return nil on frame == nil
If the frame is nil at this point, then the cause of this is likely
due to chrome not sending a frameAttached event ahead of time. This
isn't a bug in chrome, and seems to be intended behavior. Instead
of worrying about the nil frame and causing the test to fail when
the frame is nil, we can instead return early. The frame will
be initialized when getFrameTree CDP request is made, which will
call onFrameAttached and onFrameNavigated.