browser: disable a bunch of tests on windows
It seems like quite a lot of tests are not passing on windows.

Unfortunately given the volume fixing them right now seems unlikely.
So I opened https://github.com/grafana/k6/issues/4118

This tests have not been ran on windows it seems ever, before the move
of the codebase to core.