Warn on shorthand options overriding scenarios
This is currently implemented only for scenarios as that is the case
that is most confusing. Implementing this for duration overriding stages
(for example) is possible, but it seems to not be nearly hit enough.

Unfortunately due to too many of things of #883 and the way shorthand
options are handled in particular, the code can't be in any Apply
function with enough information to have a nice message.

Fixes #2529