Basic Secret Source Implementation
Add extensible Secret Sources that can be used to get secrets that will
be redacted from logs.

The in core secret sources are a file and mock based one. One is reading
from key=value file, the other takes secrets as part of its argument.

This likely will be extended in the future with more secure ways to do
secrets.

All secret values are being redacted from the logs of k6 before they go
anywhere. Removing them from other places is not in scope.

Closes #4139