Fix dispose of original handle
We don't want to dispose the new handle, only the original one, and
only when the handle is successfully adopted.

We also don't want to end the iteration if we fail to dispose of the
original handle since we're not going to be working with it. Chances
are that chrome will have dealt with it already.