Fix fetchBody with a retry
There seems to be a race condition with chrome and CDP requests when
trying to retrieve the response body. If the call to retrieve the
response body is too "quick", then we get an error back.

However with retries and a sleep it seems to solve the issue with the
small tests i have performed. I suspect this will be a greater issue in
the wild.