browser: Mark a bunch of places where races happen
In a lot of places in the k6ext.Promise will create a separate goroutine
and then access the arguments in it. This isn't safe and will race with
anything running on the event loop.

It is very likely that this happens very rarely especially as in a bunch
of the cases this is only for `opts` - which seems to be less used in
most cases.

Additionally in a bunch of cases mapping back to the sobek.Values is
done in that same goroutine instead of on the event loop.

I am marking all of those in one commit and not fixing them all as that
will take a lot longer and likely a bunch of them will not be the same
as the rest.