Fix E2E browser CI flakiness by reducing example number of browser
It seems like on MacOS running 3 browsers is way too much and it
regularly fails.

Given that the failure is due to threshold being blown over as the CI
just can't load pages fast enough, I see no point for this to keep being
flaky.

The example is about multiple scenarios not about multiple browsers in
scenarios as well.