browser: fix some mappings and remove overzealous TODOs
While trying to fix some mappings that originally I recognized as places
that races are possible, I found that is not the case.

As mappings are just a map that is methods/fields to their values, as
long as no usage the runtime happens outside of any mapping it is fine.

This is the case for most cases, this fixes the one that isn't.