Try to make TestConstantArrivalRateRunCorrectRate less flaky
The test is testing how constant arrival rate is starting iterations
over time. But it used time.Sleep of 1 second which means that it
doesn't actually check every second it checks and then it waits for 1
second and then again.

This seems to move enough, especially on windows, that it sometimes
fails.

This PR uses time.Ticker to make it check every second, making it more
likely that it won't drift.