Add `page.on('request')` (#4290)
* Add request to the PageOnEvent type
This is where the request will be set for the page.on handler to read.

* Add a new page on request event

* Add a onRequest method

This method is to be called when a new request is about to be sent
from chrome to the WuT. It takes the request and send it to the page.on
handler where the user test script can read the request data from.

* Add call to onRequest from network manager

* Update the page on mapping with the request event

* Refactor metricInterceptor to eventInterceptor

* Fix request.Timing

This is a temporary fix.

When working with page.on('request') the timing values haven't yet been
received, which are part of the response object. This will need to be
fixed later when we can wait for a response while working with
page.on('request').

* Add a test for page.on('request')

* Remove call to Body.Close() from server

* Refactor test to work with tb.url

* Refactor test to reduce boilerplate code

The issue was that HeadersArray was out of order. If they're put in
order then the comparison can be made.

* Refactor test to work with slices.IndexFunc