httpext: replace HTTP digest library
The current HTTP digest library,
`github.com/Soontao/goHttpDigestClient`, has not been updated since
March 2017 and is known to be buggy.

This commit replaces it with `github.com/icholy/digest`, a more modern
library that provides an `http.RoundTripper` implementation, allowing
digest challenges to be reused. This aligns well with our use case.

Closes https://github.com/grafana/k6/issues/800.

Signed-off-by: Eng Zer Jun <engzerjun@gmail.com>