Fix options parsing error printing the most of the options
The original idea of #4602 was to report at least the key with
potentially the option in question. Due to the lack of support from the
error - it can't parse it. The commit was trying to get everything
between the previous and next new line.

But in practice was getting anything until the end, both because of
error in the coed to get the index, but also because there were no
newlines due to how we marshal the data before that.

This should fix both and make it less likely we just get everything
until the end.