WebCrypto: fix its errors not propagating from the tests
Originally one of the edits that was done to the test would mean that
any exception would propagate and stop the whole execution.

This was later reverted in
https://github.com/grafana/xk6-webcrypto/commit/03d70232100cc657e1564a51039e049d1775421e
After this change every exception would be eaten by this code and not
propagate upwards.

Upstream(the tests) this was done so different runners have easier time
running the tests.

And after reading testharness.js docs for a while it is obvious that you
are meant to get the result in another way which is now being done in
this PR.

This probably can be done better and we can similarly to tc39 tests
record breaking ones instead of editing files, but this should be done
in a separate patch.