k6/browser: Fix race conditions from stringifying types
Before this in cases like #4564 you can have an error that stringifies
ElementHandle object. That object has many other objects inside of it,
some of which have maps or other more complex structure. None of which
are thread safe to be accessed without locking.

This is very likely the reason for the race conditions in #4564 (maybe
other cases as well). In this case we do have ElementHandle as the
argument - we get an error due to likely it getting disposed while
converting it. And then we start returning an error that has it just
stringified. While fmt goes through the object we get another event
(maybe exactly the same) modifying one of the many maps within the
object - likely under the Frame specifically.

This also likely leads to pretty useless logs due to the amount of
fields both have. With this change we only logs some fields that are
less likely to be racy, or at least won't lead to a runtime race
condition without the race detector being on.

Fixes #4564