Fix user in dockerfile
This fix helps those who are working with the image in k8s. Before this
change the user had to work with runAsUser in the pod manifest file,
which could lead to issues if the correct user wasn't used (which is
only 12345). This will mitigate that risk since we default to working
with 12345 and not k6.