Add Prometheus metrics endpoint
I have a friend who is using k6 for some massive database load-tests
that run for hours. However, k6 is OOMing sometimes and they'd like to
understand if they need to modify GOGC values.

The easiest way for them to do it is to leverage the Prometheus Golang
runtime metrics. This was brought up before in #3535 however, it was
rejected for two reasons:

1. `client_golang` dependency might go away in favor of OTel
2. The future of metrics exposition is OTel

However I'd like to reopen this for two reasons. `client_golang` is a
dependency of `opentelemetry-go` as well. So that dependency is not going
away.

But the more important reason is that there are already many high
performance Golang systems exposing `/metrics` and the metrics and tooling
around them is pretty mature and extensive.

The OTel Golang runtime metrics are new and not yet stable: https://opentelemetry.io/docs/specs/semconv/runtime/go-metrics/
Reimplementing everything for these metrics is not a trivial effort.

Signed-off-by: gouthamve <gouthamve@gmail.com>