chore: move to enumer versions that are working
The previous enumer implementation doesn't work with go > 1.20 which
practically makes it unusable. It has also have not been updated in
years.

There is a fork that unfortunately adds some additional functionality so
I decided to fork and bump versions.

But we already use this in a couple of places.

At this point I will make certain that the current code works as before
and we can agree on whether we use my minimal fork or the other one with
the additional (forced) functionality.