ci/deps: fail on go generate failing + refactor
Move the whole generation part in the Makefile and pin everything used
in it. Including protoc, which will only work on linux or WSL, maybe.

This has to do some ... updates to some of the protobuf generated files
as they are now build with a specific version.

All versions are the highest one used so far, but no more recent to
decrease need to change stuff.