Fix coordinates
When an element belongs to an iframe we need to take into account the
position of the iframe to correctly work out the coordinates for the
page.

We also want to check the hit target against the iframe and not the
element as it causes an issue with finding hit targets in the injected
script. This is the same behaviour in Playwright (using the iframe in
checkHitTarget).