Add a iframe CORS check in inject_script
This will be used to check if the iframe can access its parent. If it
can then we don't need to worry about CORS, and so when we retrieve the
click position for an element in an iframe, the position will already
be relative to the page.

If it returns false we need to remember to take the position of the
click relative to the page not the iframe.