Update translate point logic
When getting the boudingbox of a element that we want to click on, in
most cases the point is the page coords space, even when the element is
in an iframe from a different origin (CORS). However, in other cases,
the boundingBox is relative to its parents frame and not to the page.
The new logic in translate point takes that into account. CORS doesn't
seem to be the issue.