Fix issue with taskqueue and waitForNavigation
This is a temporary fix on waitForNavigation. When the taskqueue is
created, it seems that we need to perform a manual cleanup with
page.close. Since this is a common API, the first step is to mitigate
the deadlock once the iteration ends and document that page.close must
be called when working with the URL option on waitForNavigation.