Fix getting element description
If the element is in an iframe, then using the page's session will not
contain the element that we're looking for since frames do not
share context between each other.

Instead we use the element's session to retrieve the description of
itself, which works even when we're handling elements that are not in
an iframe.